Se agrego logica de archivos base de repository, service y controller

+++ CatalogService/Controllers/BaseController.cs
@@ -2,64 +2,102 @@
 using CatalogService.Infrastructure.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace CatalogService.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public abstract class BaseController<T> : ControllerBase where T : BaseEntity
+    public abstract class BaseController<T, TDto, TCreateDto> : ControllerBase
+        where T : class
     {
-        protected readonly IGenericService<T> _service;
-        protected BaseController(IGenericService<T> service)
+        protected readonly IGenericService<T, TDto, TCreateDto> _service;
+
+        public BaseController(IGenericService<T, TDto, TCreateDto> service)
         {
             _service = service;
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public virtual async Task<IActionResult> GetAll()
         {
-            var entities = await _service.GetAllAsync();
-            return Ok(entities);
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim == null)
+                return Unauthorized();
+
+            var dtos = await _service.GetAllAsync();
+            return Ok(dtos);
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetById(int id)
+        public virtual async Task<IActionResult> GetById(int id)
         {
-            var entity = await _service.GetByIdAsync(id);
-            if (entity == null)  return NotFound();
-            return Ok(entity);
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim == null)
+                return Unauthorized();
+
+            var dto = await _service.GetByIdAsync(id);
+            if (dto == null)
+                return NotFound();
+
+            return Ok(dto);
         }
 
         [HttpPost]
-        public async Task<IActionResult> Create(T entity)
+        public virtual async Task<IActionResult> Create(TCreateDto model)
         {
-            await _service.AddAsync(entity);
-            return CreatedAtAction(nameof(GetById), new { id = entity.Id }, entity);
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim == null)
+                return Unauthorized();
+
+            var validationError = await _service.ValidateBeforeSave(model);
+            if (validationError != null) return BadRequest(validationError);
+
+            var createdEntity = await _service.CreateAsync(model);
+            return Ok(createdEntity);
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Update(int id, T entity)
+        public virtual async Task<IActionResult> Update(int id, TCreateDto model)
         {
-            if (id != entity.Id) return BadRequest();
-            await _service.UpdateAsync(entity);
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim == null)
+                return Unauthorized();
+
+            var validationError = await _service.ValidateBeforeSave(model);
+            if (validationError != null) return BadRequest(validationError);
+
+            var updatedDto = await _service.UpdateAsync(id, model);
+            if (updatedDto == null)
+                return NotFound();
+
+            return Ok(updatedDto);
+        }
+
+        [HttpPut("{id}/status")]
+        public virtual async Task<IActionResult> UpdateStatus(int id, bool isActive)
+        {
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim == null)
+                return Unauthorized();
+            var success = await _service.UpdateStatusAsync(id, isActive);
+            if (!success)
+                return NotFound();
             return NoContent();
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(int id)
+        public virtual async Task<IActionResult> Delete(int id)
         {
-            await _service.DeleteAsync(id);
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            if (userIdClaim == null)
+                return Unauthorized();
+
+            var success = await _service.DeleteAsync(id);
+            if (!success)
+                return NotFound();
+
             return NoContent();
         }
-
-        [HttpPatch("{id}/archive")]
-        public async Task<IActionResult> Archive(int id)
-        {
-            await _service.ArchiveAsync(id);
-            return NoContent();
-        }
-
-
-
     }
 }
